+++ doc/releases/changelog-dev.md
@@ -228,8 +228,9 @@ such as `shots`, `rng` and `prng_key`.
 
 <h4>Capturing and representing hybrid programs</h4>
 
-* Functions and plxpr can now be natively transformed using the new `qml.transforms.DecomposeInterpreter`.
-  This class decomposes pennylane operators following the same API as `qml.transforms.decompose`.
+* Functions and plxpr can now be natively transformed using the new `qml.capture.transforms.DecomposeInterpreter`
+  when program capture is enabled. This class decomposes pennylane operators following the same API as
+  `qml.transforms.decompose`.
   [(#6691)](https://github.com/PennyLaneAI/pennylane/pull/6691)
 
 * Execution with capture enabled now follows a new execution pipeline and natively passes the
