+++ pennylane/transforms/__init__.py
@@ -344,4 +344,4 @@ from .tape_expand import (
 from .transpile import transpile
 from .zx import to_zx, from_zx
 from .broadcast_expand import broadcast_expand
-from .decompose import decompose, DecomposeInterpreter
+from .decompose import decompose
